+++ src/program/lwaftr/tests/propbased/genyang.lua
@@ -6,10 +6,10 @@ module(..., package.seeall)
 local ffi    = require("ffi")
 local schema = require("lib.yang.schema")
 
-local capabilities = {['ietf-softwire']={feature={'binding', 'br'}}}
+local capabilities = {['ietf-softwire-br']={feature={'binding'}},}
 require('lib.yang.schema').set_default_capabilities(capabilities)
 
-local schemas = { "ietf-softwire", "snabb-softwire-v1" }
+local schemas = { "ietf-softwire-br", "snabb-softwire-v2" }
 
 -- toggles whether functions should intentionally generate invalid
 -- values for fuzzing purposes